Allow custom environments
I'm currently working on an application that has an extra environment
for demonstrations. It is another Heroku instance running alongside our
production and staging environments. It shares staging behavior in that
we will occasionally copy production data over to it.

This commit makes staging a pre-defined case of a Heroku environment.
Users can make their own app binstubs for custom environments such as
"beta", "demo", etc. that can take advantage of parity conveniences.