Enable redis-cli access to RedisToGo
This commit helps accomplish the functionality described in:

https://github.com/croaky/parity/issues/8

Using `Open3` from the standard library to capture
the output of the request to Heroku for the URL,
which is then memoized and parsed with regex to get
the host, post, and password.

User can then access their Redis service with `production redis_cli`.

Make `REDIS_URL` default environment variable for Redis config.