Release binaries using Travelling Ruby
Distributing command-line applications as Ruby gems is cumbersome. It
requires users to reinstall the application every time they upgrade
their Ruby version and it breaks on machines that run older versions of
Ruby. It also makes installation more painful for non-Ruby devs.

A better approach is to use a package manager such as Homebrew. This PR
creates a set of `rake` tasks to generate packages that can be tied to a
release. It also bundles [traveling ruby] so as not to depend on the
existing rubies on a user's system.

The packages contain shims for Parity's binaries which run them
using Traveling Ruby. The generated packages have the following
structure:

```
parity-0.4.1-osx
├── bin # shims
│   ├── development
│   ├── production
│   └── staging
└── lib
    ├── app # parity app
    │   ├── README.md
    │   ├── bin
    │   └── lib
    └── ruby # traveling ruby
```

Packages can be generated with rubies for the following targets:

* OSX `rake package:osx`
* Linux x86 `rake package:linux:x86`
* Linux x86_64 `rake package:linux:x86_64`

or for all packages `rake package:all`

These packages would then be attached to a [release] on github. They
could be manually downloaded and extracted to a directory with `/bin`
was in `PATH` or they could be installed via a package manager.

I created a [working Homebrew formula][formula] that pulls
[this release] on my fork as a proof of concept.

[release]: https://github.com/croaky/parity/releases
[formula]: https://github.com/joelq/homebrew-formulae/blob/master/Formula/parity.rb
[this release]: https://github.com/JoelQ/parity/releases/tag/v0.4.3
[traveling ruby]: https://github.com/phusion/traveling-ruby