Use new pg:backups syntax
We attempted to do this in 5e789c3, but Heroku Toolbelt had some missing
functionality with respect to the new `pg:backups` syntax, and our code
retained the `--expire` flag, which is now automatic and no longer
allowed.

This change requires that users have version 3.32.0 of the Heroku
Toolbelt installed, but the old `pgbackups:capture / pgbackups:restore`
syntax and addon are being retired in the first week of May 2015, so we
need to do this to move on in any event (users who don't upgrade won't
be able to use the old parity either).

Close #29.