Allow bypass of confirmation when restoring to staging
Heroku asks for confirmation with `--confirm appname` when performing
potentially destructive CLI operations such as a database restore. If
staging data is tracking with production or easily rebuilt using seeds,
this confirmation can be annoying, particularly during the common
operation of taking a snapshot of production and then immediately
copying it to staging to sync staging's DB up with production's.

This commit changes the behavior of non-production, non-local
development environments to append the confirmation argument on database
restores to overwrite the staging or other feature environment database
when restored from another environment.

Close #38.