Fix `deploy` problems in 0.9.0
v0.9.0 broke `deploy` in two ways. The first is that migrations were
being checked after the deploy completed, which resulted in the check
for diffs in `db/migrate` always finding no differences. This change
reverts to the prior pattern of caching whether migrations are pending
at the start of `deploy` before pushing to Heroku.

The second is that the check for `rake` would raise an exception if a
user's local bundle wasn't installed, as the shell command would return
a failing status. `Kernel.system` can return `true`, `false`, or `nil`,
but `Kernel.exit` can't accept `nil` as an argument.
`Parity::Environment#run` will now always return a boolean value so that
`Kernel.exit` receives an appropriate argument.

Fix #61.