Use a shorter check for a Rails app on deploy
In 17448e9 we added a check to determine if the application was running
Rails (specifically whether the app had access to Rake and a
`db:migrate` task) through Rake's "dry-run" functionality. This approach
produces some noisy output and adds a non-trivial amount of time to the
deploy process.

Rather than start Rake and execute the dry run on every deploy, we'll
instead check for a `Rakefile` and the presence of `db/migrate` in the
application's folder structure, which will take less time and produce
less noise in `STDOUT`.