Update docs on deploying non-master branches
In df296fe6e77698fe4522b120944621667b1dcbd3 we added the capability to
push non-master branches to staging or other feature environments for
testing. Pushes to non-production environments are forced in case a
developer has been squashing commits.

This change updates the README to clarify that pushes are permitted from
non-master branches to non-production environments.

Close #56.