Parse git remote name to determine Heroku app
Several reports have matched #72, which was affecting the `restore`
command for some users. The cause turned out to be local folder names
not matching the name of the Heroku app. Now that Heroku tracks Git
remotes, we can take advantage of the Git remote to correctly identify
the app name even the local folder name doesn't match.

While this requirement has been present for the entirety of Parity's
existence, it stopped being a problem for most commands once Heroku
introduced remotes. I inadvertently reintroduced this requirement in
ad2c21a42fb85bd00ec58d7326725f0a1d5a7771 when we started automatically
bypassing confirmation arguments for restores to non-production
environments.

I've moved the `git` and `rake` dependencies to the gemspec, as they do
not appear to be installed by default if they are listed in the Gemfile.
Fix #72.