Parse database.yml that uses ERB
Rails allows the use of ERB to define dynamic values in `database.yml`.
If the database name in development is defined using a dynamic value,
the ERB syntax is passed in as an argument to Heroku-CLI rather than the
desired name.

This change parses `database.yml` as an ERB template, reading in any
development variables from the `.env` file as needed.

Other changes:
- development-only gems are sectioned off into a separate group in the
  Gemfile

Close #73.