Update packaging with bundled dependencies
Version 0.9.3 will introduce dependencies on the `dotenv` and `git`
gems. This will be the first time that we include gem dependencies in
our Travelling Ruby distribution.

The included changes update the packaging script to include vendored gem
dependencies and the necessary `Gemfile`/`Gemfile.lock` files for
Travelling Ruby to find when the user runs `development`, `staging`, or
`production`. The Travelling Ruby shims are updated to find the included
gem manifest and vendored code.

At the moment, the packaging task requires manual intervention, as
Travelling Ruby uses Bundler v1.9.9, which attempts to modify
`Gemfile.lock` files built by versions of Bundler from 1.10 and later.
When phusion/traveling-ruby#81 is merged, we will be able to use our
existing `Gemfile.lock` without temporary editing to build dependencies
and bundle them with a temporarily modified Gemfile.lock.

Thanks to @tabfugnic for helping resolve the dependency location issues!