Packaging: Remove Bundler from packaging
Traveling Ruby is no longer receiving updates, and is locked on a
version of Bundler that causes an update to `Gemfile.lock` when `bundle
install` is run on installed systems, which causes permission problems
on client users' machines.

We stopped using gem dependencies in Parity in 5704201, but still had
pieces of our packaging that were there solely to support gem
dependencies.

This change brings our Rakefile back to the original setup for Traveling
Ruby by @joelq in 7916985 along with @gabebw's shim fixes in 89be4ad. We
still plan to migrate away from Traveling Ruby, but these changes are
needed to get a valid, working packaged version of Parity on client
machines to comply with the new Heroku CLI format.

One update has been made since those prior Rakefile and shim versions,
which is to update Traveling Ruby to use version 2.2.2 as of 2015-07-15,
as it contains security vulnerability fixes not present in prior
versions.