Remove Traveling Ruby
Traveling Ruby is stuck on a version of Bundler that has a bug that
prevents us from packaging `Gemfile.lock` without needing to touch it on
the installed machine's filesystem. There hasn't been a commit in almost
ten months, and it appears that development has stopped.

We're going to try packaging Parity directly and using system Ruby
instead of bundling our own Ruby executable. This change removes all of
the packaging tooling related to Traveling Ruby.