Heroku-CLI change: don't pass `--confirm` to `pg:push`
Heroku's `pg:push` CLI no longer accepts the `--confirm` argument, as it
expects the remote database to already be reset before overwriting it
with a backup.

Before restoring to a remote environment, we reset the remote database
with `pg:reset`. We will pass a confirm argument to `pg:reset`. To
accomplish this, the mechanism to determine the remote application name
(not the git remote) has been extracted into its own object.

Fix #128.