development restore: ensure a tmp directory exists
When restoring a remote backup to development, if an application doesn't
have a `tmp` folder (either because it wasn't in source control or the
app never has one), we can't download the backup file from Heroku.

This change uses `mkdir -p` to create the directory (if needed) before
downloading the backup to the local machine.

Close #130.