Allow pre-2.2 Rubies to execute `development restore`
In #143, we got a report that because `Etc.nprocessors` was not
implemented until Ruby 2.2 that Ruby 2.1 users were unable to execute
`development restore`, which leverages that method to determine how many
parallel processes to use when restoring the backup from the downloaded
dump.

This change implements a check to see if `Etc.nprocessors` is available.
In the event that it's not, we'll default to two parallel processes. We
will plan to sunset this change in July 2018, giving users on Ruby 2.1,
which has been EOLed, six months to move on.

Close #143.