Drop ar_internal_metadata table on local restore if present
This change updates the local development restore operation to drop the
`ar_internal_metadata` table if it's present. This table is used by
Rails 5.x to keep track of the environment, but does not need to be
present for development. If it is present, users are unable to drop
their local database with Rake due to a Rails check that aims to protect
users from inadvertently dropping their production database.

Close #147