Use different approach for marking `ar_internal_metadata` on restore
The previous attempt in e5e025a to mark the `ar_internal_metadata`
environment row as non-production does not work reliably. The change
here fixes the issue with a slightly more brute-force approach, but one
that does not require capabilities beyond pure SQL.

If the `ar_internal_metadata` table does not exist, we'll create a blank
one with no rows. The following update operation will no-op.