Use bundle exec when starting remote console
The combination of Bundler 2, Heroku's Ruby buildpack, and Rails' own
binstubs for Rails 5.2 appears to cause a conflict when starting the
console for an application bundled with Bundler 2.x.

```
% production console
Running rails console on ⬢ app-production... up, run.5165 (Standard-1X)
You must use Bundler 2 or greater with this lockfile.
```

The solution is to invoke the console with `bundle exec`, which gets
around the mismatch of Bundler 1.x and 2.x and blocks the Rails console
from starting.