Drop non-Rubygem installs
Since Travelling Ruby is no longer supported, it's no longer feasible
for us to package a self-contained release of Parity that runs
independently of local Rubies. This change acknowledges our trailing
support of Homebrew- and apt-installed Parity and encourages
installation of the Rubygem version instead.

We will remove for formula from thoughtbot's Homebrew formulae in a
subsequent change.