Fix specs for new default branch, bump development Ruby -> 2.7.2
Since the default branch became `main`, some specs that relied on the
default being `master` were failing. This change adds setup for examples
that depend on the old `master` branch being the default branch so that
the specs continue to work for both cases.

Bump the development version of Ruby to 2.7.2.